Move class definitions outside packages
