+++ scio-elasticsearch/common/src/main/scala/com/spotify/scio/elasticsearch/syntax/SCollectionSyntax.scala
@@ -1,0 +1,69 @@
+/*
+ * Copyright 2023 Spotify AB
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.spotify.scio.elasticsearch.syntax
+
+import co.elastic.clients.elasticsearch.core.bulk.BulkOperation
+import com.spotify.scio.elasticsearch.ElasticsearchIO.{RetryConfig, WriteParam}
+import com.spotify.scio.elasticsearch.{ElasticsearchIO, ElasticsearchOptions}
+import com.spotify.scio.io.ClosedTap
+import com.spotify.scio.values.SCollection
+import org.apache.beam.sdk.io.elasticsearch.ElasticsearchIO.Write.BulkExecutionException
+import org.joda.time.Duration
+
+class ElasticsearchSCollectionOps[T](@transient private val self: SCollection[T]) extends AnyVal {
+
+  /**
+   * Save this SCollection into Elasticsearch.
+   *
+   * @param esOptions
+   *   Elasticsearch options
+   * @param flushInterval
+   *   delays to Elasticsearch writes for rate limiting purpose
+   * @param f
+   *   function to transform arbitrary type T to Elasticsearch `DocWriteRequest`
+   * @param numOfShards
+   *   number of parallel writes to be performed, recommended setting is the number of pipeline
+   *   workers
+   * @param errorFn
+   *   function to handle error when performing Elasticsearch bulk writes
+   */
+  def saveAsElasticsearch(
+    esOptions: ElasticsearchOptions,
+    flushInterval: Duration = WriteParam.DefaultFlushInterval,
+    numOfShards: Long = WriteParam.DefaultNumShards,
+    maxBulkRequestOperations: Int = WriteParam.DefaultMaxBulkRequestOperations,
+    maxBulkRequestBytes: Long = WriteParam.DefaultMaxBulkRequestBytes,
+    errorFn: BulkExecutionException => Unit = WriteParam.DefaultErrorFn,
+    retry: RetryConfig = WriteParam.DefaultRetryConfig
+  )(f: T => Iterable[BulkOperation]): ClosedTap[Nothing] = {
+    val param = WriteParam(
+      f,
+      errorFn,
+      flushInterval,
+      numOfShards,
+      maxBulkRequestOperations,
+      maxBulkRequestBytes,
+      retry
+    )
+    self.write(ElasticsearchIO[T](esOptions))(param)
+  }
+}
+
+trait SCollectionSyntax {
+  implicit def elasticsearchSCollectionOps[T](sc: SCollection[T]): ElasticsearchSCollectionOps[T] =
+    new ElasticsearchSCollectionOps[T](sc)
+}
