+++ scio-extra/src/main/scala/com/spotify/scio/extra/annoy/package.scala
@@ -17,17 +17,6 @@
 
 package com.spotify.scio.extra
 
-import java.util.UUID
-
-import com.spotify.scio.ScioContext
-import com.spotify.scio.annotations.experimental
-import com.spotify.scio.values.{SCollection, SideInput}
-import org.apache.beam.sdk.transforms.{DoFn, View}
-import org.apache.beam.sdk.values.PCollectionView
-import org.slf4j.LoggerFactory
-
-import scala.jdk.CollectionConverters._
-
 /**
  * Main package for Annoy side input APIs. Import all.
  *
@@ -93,184 +82,12 @@ import scala.jdk.CollectionConverters._
  *   }
  * }}}
  */
-package object annoy {
-  sealed abstract class AnnoyMetric
-  case object Angular extends AnnoyMetric
-  case object Euclidean extends AnnoyMetric
+package object annoy extends syntax.AllSyntax {
 
-  private val logger = LoggerFactory.getLogger(this.getClass)
-
-  /**
-   * AnnoyReader class for approximate nearest neighbor lookups. Supports vector lookup by item as
-   * well as nearest neighbor lookup by vector.
-   *
-   * @param path
-   *   Can be either a local file or a GCS location e.g. gs://<bucket>/<path>
-   * @param metric
-   *   One of Angular (cosine distance) or Euclidean
-   * @param dim
-   *   Number of dimensions in vectors
-   */
-  class AnnoyReader private[annoy] (path: String, metric: AnnoyMetric, dim: Int) {
-    require(dim > 0, "Vector dimension should be > 0")
-
-    import com.spotify.annoy._
-
-    private val index = {
-      val indexType = metric match {
-        case Angular   => IndexType.ANGULAR
-        case Euclidean => IndexType.EUCLIDEAN
-      }
-      new ANNIndex(dim, path, indexType)
-    }
-
-    /** Gets vector associated with item i. */
-    def getItemVector(i: Int): Array[Float] = index.getItemVector(i)
-
-    /** Copies vector associated with item i into vector v. */
-    def getItemVector(i: Int, v: Array[Float]): Unit = index.getItemVector(i, v)
-
-    /** Gets maxNumResults nearest neighbors for vector v. */
-    def getNearest(v: Array[Float], maxNumResults: Int): Seq[Int] =
-      index.getNearest(v, maxNumResults).asScala.toSeq.asInstanceOf[Seq[Int]]
-  }
-
-  /** Enhanced version of [[ScioContext]] with Annoy methods. */
-  implicit class AnnoyScioContext(private val self: ScioContext) extends AnyVal {
-
-    /**
-     * Create a SideInput of [[AnnoyReader]] from an [[AnnoyUri]] base path, to be used with
-     * [[com.spotify.scio.values.SCollection.withSideInputs SCollection.withSideInputs]]
-     *
-     * @param metric
-     *   Metric (Angular, Euclidean) used to build the Annoy index
-     * @param dim
-     *   Number of dimensions in vectors used to build the Annoy index
-     */
-    @experimental
-    def annoySideInput(path: String, metric: AnnoyMetric, dim: Int): SideInput[AnnoyReader] = {
-      val uri = AnnoyUri(path, self.options)
-      val view = self.parallelize(Seq(uri)).applyInternal(View.asSingleton())
-      new AnnoySideInput(view, metric, dim)
-    }
-  }
-
-  implicit class AnnoyPairSCollection(@transient private val self: SCollection[(Int, Array[Float])])
-      extends AnyVal {
-
-    /**
-     * Write the key-value pairs of this SCollection as an Annoy file to a specific location,
-     * building the trees in the index according to the parameters provided.
-     *
-     * @param path
-     *   Can be either a local file or a GCS location e.g. gs://<bucket>/<path>
-     * @param metric
-     *   One of Angular (cosine distance) or Euclidean
-     * @param dim
-     *   Number of dimensions in vectors
-     * @param nTrees
-     *   Number of trees to build. More trees means more precision & bigger indices. If nTrees is
-     *   set to -1, the trees will automatically be built in such a way that they will take at most
-     *   2x the memory of the vectors.
-     * @return
-     *   A singleton SCollection containing the [[AnnoyUri]] of the saved files
-     */
-    @experimental
-    def asAnnoy(path: String, metric: AnnoyMetric, dim: Int, nTrees: Int): SCollection[AnnoyUri] = {
-      val uri = AnnoyUri(path, self.context.options)
-      require(!uri.exists, s"Annoy URI ${uri.path} already exists")
-
-      self.transform { in =>
-        in.groupBy(_ => ())
-          .map { case (_, xs) =>
-            logger.info(s"Saving as Annoy: $uri")
-            val startTime = System.nanoTime()
-            val annoyWriter = new AnnoyWriter(metric, dim, nTrees)
-            try {
-              val it = xs.iterator
-              while (it.hasNext) {
-                val (k, v) = it.next()
-                annoyWriter.addItem(k, v)
-              }
-              val size = annoyWriter.size
-              uri.saveAndClose(annoyWriter)
-              val elapsedTime = (System.nanoTime() - startTime) / 1000000000.0
-              logger.info(s"Built index with $size items in $elapsedTime seconds")
-            } catch {
-              case e: Throwable =>
-                annoyWriter.free()
-                throw e
-            }
-            uri
-          }
-      }
-    }
-
-    /**
-     * Write the key-value pairs of this SCollection as an Annoy file to a temporary location,
-     * building the trees in the index according to the parameters provided.
-     *
-     * @param nTrees
-     *   Number of trees to build. More trees means more precision & bigger indices. If nTrees is
-     *   set to -1, the trees will automatically be built in such a way that they will take at most
-     *   2x the memory of the vectors.
-     * @return
-     *   A singleton SCollection containing the [[AnnoyUri]] of the saved files
-     */
-    @experimental
-    def asAnnoy(metric: AnnoyMetric, dim: Int, nTrees: Int): SCollection[AnnoyUri] = {
-      val uuid = UUID.randomUUID()
-      val tempLocation = self.context.options.getTempLocation
-      require(tempLocation != null, s"--tempLocation arg is required")
-      val path = s"$tempLocation/annoy-build-$uuid"
-      this.asAnnoy(path, metric, dim, nTrees)
-    }
-
-    /**
-     * Write the key-value pairs of this SCollection as an Annoy file to a temporary location,
-     * building the trees in the index according to the parameters provided, then load the trees as
-     * a side input.
-     *
-     * @param metric
-     *   One of Angular (cosine distance) or Euclidean
-     * @param dim
-     *   Number of dimensions in vectors
-     * @param nTrees
-     *   Number of trees to build. More trees means more precision & bigger indices. If nTrees is
-     *   set to -1, the trees will automatically be built in such a way that they will take at most
-     *   2x the memory of the vectors.
-     * @return
-     *   SideInput[AnnoyReader]
-     */
-    @experimental
-    def asAnnoySideInput(metric: AnnoyMetric, dim: Int, nTrees: Int): SideInput[AnnoyReader] =
-      self.asAnnoy(metric, dim, nTrees).asAnnoySideInput(metric, dim)
-  }
-
-  /** Enhanced version of [[com.spotify.scio.values.SCollection SCollection]] with Annoy methods */
-  implicit class AnnoySCollection(@transient private val self: SCollection[AnnoyUri])
-      extends AnyVal {
-
-    /**
-     * Load Annoy index stored at [[AnnoyUri]] in this
-     * [[com.spotify.scio.values.SCollection SCollection]].
-     * @param metric
-     *   Metric (Angular, Euclidean) used to build the Annoy index
-     * @param dim
-     *   Number of dimensions in vectors used to build the Annoy index
-     * @return
-     *   SideInput[AnnoyReader]
-     */
-    @experimental
-    def asAnnoySideInput(metric: AnnoyMetric, dim: Int): SideInput[AnnoyReader] = {
-      val view = self.applyInternal(View.asSingleton())
-      new AnnoySideInput(view, metric, dim)
-    }
-  }
-
-  private class AnnoySideInput(val view: PCollectionView[AnnoyUri], metric: AnnoyMetric, dim: Int)
-      extends SideInput[AnnoyReader] {
-    override def get[I, O](context: DoFn[I, O]#ProcessContext): AnnoyReader =
-      context.sideInput(view).getReader(metric, dim)
-  }
+  @deprecated("Use syntax.AnnoyScioContextOps instead", "0.14.0")
+  type AnnoyScioContext = syntax.AnnoyScioContextOps
+  @deprecated("Use syntax.AnnoyPairSCollectionOps instead", "0.14.0")
+  type AnnoyPairSCollection = syntax.AnnoyPairSCollectionOps
+  @deprecated("Use syntax.AnnoySCollectionOps instead", "0.14.0")
+  type AnnoySCollection = syntax.AnnoySCollectionOps
 }
