+++ scio-extra/src/main/scala/com/spotify/scio/extra/sparkey/LargeHashSCollectionFunctions.scala
@@ -18,10 +18,12 @@
 package com.spotify.scio.extra.sparkey
 
 import com.spotify.scio.extra.sparkey.instances.SparkeySet
+import com.spotify.scio.extra.sparkey.syntax.SCollectionSyntax
 import com.spotify.scio.values.{SCollection, SideInput}
 
 /** Extra functions available on SCollections for Sparkey hash-based filtering. */
-class LargeHashSCollectionFunctions[T](private val self: SCollection[T]) {
+class LargeHashSCollectionFunctions[T](private val self: SCollection[T]) extends AnyVal {
+  import SCollectionSyntax._
 
   /**
    * Return a new SCollection containing only elements that also exist in the `LargeSetSideInput`.
